requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.8
python-dotenv==0.21.1
python-slugify==8.0.1
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.6.2
unicode==2.9
uvicorn==0.22.0
